windows: the three lints only a Windows runner could see — and the guard for them
Real Windows, first engine run: build ok, stage 1 ok, stage 2 ok, **clippy
failed**. Three lints, all inside `cfg(windows)` code that the lint pass on a
Linux host never compiles:

- `create(true)` without `truncate` on the writer-lock sidecar. Harmless as
  written and worth being explicit about anyway: the sidecar's CONTENT is never
  read — the lock lives in the kernel — but truncating an existing one would
  still be wrong, because another process may hold a lock on it right now.
- an `unneeded return` in `boot_id`'s Windows arm.
- `u32::from(u32::MAX - 1)` in the futex-wait timeout clamp.

This is the macOS debt of #107 repeating on a new platform, so the fix is the
one that generalises: `linux.yml` now runs `cargo clippy -p mpedb-core --lib
--target x86_64-pc-windows-gnu -- -D warnings` on every push. Cross-linting
costs seconds and moves the finding from a Windows-runner round trip to the
push that caused it. wasm32 stays at `check` — its arm has no lint history, and
`check` is what catches the failure mode that actually occurred there.

Verified: workspace clippy, mpedb-core tests, wasm32 check, Windows
cross-clippy, and 87 tests under Wine — all by exit code.

+++ crates/mpedb-core/src/wincompat.rs
@@ -235,7 +235,7 @@ pub fn errno() -> libc::c_int {
 /// `WaitOnAddress` — a direct futex equivalent, not a degradation. macOS has to
 /// poll here; Windows does not.
 pub fn futex_wait(word: &std::sync::atomic::AtomicU32, expected: u32, timeout: std::time::Duration) {
-    let ms = timeout.as_millis().min(u32::from(u32::MAX - 1) as u128) as u32;
+    let ms = timeout.as_millis().min((u32::MAX - 1) as u128) as u32;
     let cmp = expected;
     unsafe {
         WaitOnAddress(
